Minor Fixes
There were some namespace references missing, which worked by
coincidence.

Added missing header in one file.

Wrapped namespace into a single include header. This was mainly to
support including a single header when needing any of the classes. This
helps with testing, because of the circular referecences between all of
the classes and `Container`.