Make Registrar Thread Safe
Enables thread safety in the `Registrar` class which should make the
whole container thread safe. For performance reasons it uses a
read/write lock combination so that multiple reads can occur at the same
time as long as a write is not also occurring.

Adds a few more tests, including one to ensure the thread safety is
working as expected.