Add Circular Dependency Resolution Detection
Adds the ability to track circular dependencies at runtime when they are
being resolved. Attempting to reolve a service which contains a circular
dependency will throw a `std::runtime_error` when the service is
resolved.

Circular dependencies are tracked on a per-thread basis because a single
thread can only be attempting to resolve a single dependency graph at
one moment in time. This also prevents resolutions on other threads from
interfering with the detection of different threads.