Allow Passing Lambda to RegisterFactory
Previously the `RegisterFactory` method would only accept lambdas if
they were explicitly cast to `std::function` first. This is because the
compiler can't figure out when it is done expanding `TArgs...` while
attempting type-deduction. We can use `std::conditional_t<>` to give it
a hint about what the type should be.