Complete Rewrite
Rewrites cdif to use a builder pattern for constructing registrations.
Registration builders are now used directly by consumers to create a
registration. This alleviates the problem by which the `Register`
methods on `Container` were growing wildly out of hand.

Further multiple registrations are now provided for each registration
builder, these are dependent on the type of registration builder being
used and the scope in which the resulting registrations live.

Registrations now have a lifetime scope. This makes it possible to
force the container to create one instance of the class per dependency,
thread, or singleton. This exposed a flaw by which singletons in the
past were created when they were registered; now they are created when
they are first requested.

Previously `const`, `volatile`, ref, and ref-ref would cause a
resolution to fail; this has been rectified by ignoring `const` and
`volatile` in lookup and registering ref when it makes sense.