Fix Factory Registrations
Previously scoped factory registrations were broken due to misspelled
usage of a instance variable. This was fixed.

Further work was done to rework how scoped factory registrations
operate. Previously the factory function itself was scoped rather than
the return type of the factory funtion. This was a less that useful
implementation, and was changed to make the return type of the factory
the scoped object.